--- src/components/Container/index.js
@@ -1,2 +1,0 @@
-import Container from "./Container";
-export default Container;
