+++ src/components/index.js
@@ -1,4 +1,3 @@
-import Container from "./Container";
 import ScheduleRow from "./ScheduleRow";
 import ScheduleTable from "./ScheduleTable";
 import EventSchedule from "./EventSchedule";
@@ -6,14 +5,17 @@ import FAQSection from "./FAQSection";
 import AttendAndSponsorBoxes from "./AttendAndSponsorBoxes";
 import TracksAndWorkshops from "./TracksAndWorkshops";
 import Footer from "./Footer";
+import HeroText from "./HeroText";
+import FooterBar from "./FooterBar";
 
 export {
-  Container,
   ScheduleRow,
   ScheduleTable,
   EventSchedule,
   FAQSection,
   Footer,
   AttendAndSponsorBoxes,
-  TracksAndWorkshops
+  TracksAndWorkshops,
+  HeroText,
+  FooterBar
 };
